+++ PremiumScraps/Config.cs
@@ -13,7 +13,9 @@ namespace PremiumScraps
         public readonly List<ulong> unluckyPlayersID = new List<ulong>();
         public readonly List<(int, int)> scrapValues = new List<(int, int)>();
         public readonly ConfigEntry<bool> diceEvents;
+        public readonly ConfigEntry<bool> languageMode;
         public readonly ConfigEntry<string> unluckyPlayersStr;
+        public readonly ConfigEntry<bool> gazpachoMemeSfx;
         public readonly ConfigEntry<bool> squareSteelWeapon;
         public readonly ConfigEntry<bool> controllerBodyCams;
         public readonly List<ConfigEntry<int>> entries = new List<ConfigEntry<int>>();
@@ -22,10 +24,12 @@ namespace PremiumScraps
         public Config(ConfigFile cfg, List<Scrap> scraps)
         {
             cfg.SaveOnConfigSet = false;
-            diceEvents = cfg.Bind("General", "Dice events", true, "Adds some custom dice rolls to Emergency Dice items. Requires 'Emergency Dice Updated' 1.7.4+ to work, or else it will be automatically false.");
-            unluckyPlayersStr = cfg.Bind("General", "Unlucky players", "76561198984467725,76561199094139351,76561198198881967,76561198002410826", "Comma separated list of players Steam ID that you want them to be unlucky. Bad things will happen to unlucky players, use this config to take a sweet revenge on your friends...");
+            diceEvents = cfg.Bind("General", "Dice events", true, "Adds some custom dice rolls to Emergency Dice items.\nRequires 'Emergency Dice Updated' 1.7.4+ to work, or else it will be automatically false.");
+            languageMode = cfg.Bind("General", "Change Language mode", false, "This mod is automatically translated into French if your system is in french. Setting this config to 'true' will have the following effect: if YOU ARE NOT french will change the language to french, or if YOU ARE french will change the language back to english.");
+            unluckyPlayersStr = cfg.Bind("General", "Unlucky players", "76561198984467725,76561199094139351,76561198198881967,76561198002410826", "Comma separated list of players Steam ID that you want them to be unlucky.\nBad things will happen to unlucky players, use this config to take a sweet revenge on your friends...");
+            gazpachoMemeSfx = cfg.Bind("Items", "Gazpacho meme sfx", true, "Turns El Gazpacho's grab and drop sfx to memes sounds.\nWill be automatically false if the chosen language is not french.");
             squareSteelWeapon = cfg.Bind("Items", "Square Steel weapon", true, "Turns Galvanized Square Steel into a usable weapon.");
-            controllerBodyCams = cfg.Bind("Items", "Controller Body Cams", true, "Upgrade the Controller screen with a camera from OpenBodyCams. Will be automatically false is OpenBodyCams is not installed.");
+            controllerBodyCams = cfg.Bind("Items", "Controller Body Cams", true, "Upgrade the Controller screen with a camera from OpenBodyCams mod.\nWill be automatically false if OpenBodyCams is not installed.");
             foreach (Scrap scrap in scraps)
             {
                 entries.Add(cfg.Bind("Spawn chance", scrap.asset.Split("/")[0], scrap.rarity, "Rarity of the item."));